The code for two detectors (A3 and A4) has been updated to a3_spindle_detection_v02.m.txt and a4_spindle_detection_v02.m.txt.
A3
There was an error in the calculation of the background NREM signal that is used to establish the threshold.  The NREM background window was much shorter than required.  Performance of this detector (relative to the human consensus) has been improved due to this fix.

A4
Small discrepencies between the original C implementation and this implemenation have been fixed.  The changes only have a small impact on the performance (relative to the human consensus) of this detector.

The original implementations of both A3 and A4 can be found in the originalCode_2014.zip archive file.